requirements.txt
functions-framework
slack-bolt==1.18.1
google-generativeai==0.3.2
requests==2.27.1
Flask==3.0.0
gunicorn==20.1.0
Werkzeug==3.0.1
google-cloud-storage